# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API backend de la plateforme RAYCARD (wallet Mobile Money + cartes Visa virtuelles).",
        "title": "RAYCARD API",
        "contact": {
            "name": "RAYCARD — HR-Skills SARL"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/backoffice/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne les transactions tous wallets confondus, filtrables par utilisateur, statut et type.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"2. Admin - Wallet\""
                ],
                "summary": "Liste des transactions (back-office)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtre par utilisateur",
                        "name": "utilisateur_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtre par statut (en_attente, envoyee, succes, echouee)",
                        "name": "statut",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtre par type (recharge, retrait, financement_carte, annulation_carte)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wallet.TransactionDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "non authentifié",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "403": {
                        "description": "réservé aux administrateurs",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
...
        "/backoffice/wallets/{id}/degel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Réactive n'importe quel wallet gelé. Tracé dans l'audit log.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"2. Admin - Wallet\""
                ],
                "summary": "Dégel d'un wallet (back-office)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du wallet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallet.WalletDTO"
                        }
                    },
                    "401": {
                        "description": "non authentifié",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "403": {
                        "description": "réservé aux administrateurs",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "404": {
                        "description": "wallet introuvable",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "422": {
                        "description": "le wallet n'est pas gelé",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/backoffice/wallets/{id}/gel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bloque n'importe quel wallet actif : plus aucun débit ni crédit possible. Tracé dans l'audit log.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"2. Admin - Wallet\""
                ],
                "summary": "Gel d'un wallet (back-office)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du wallet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallet.WalletDTO"
                        }
                    },
                    "401": {
                        "description": "non authentifié",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "403": {
                        "description": "réservé aux administrateurs",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "404": {
                        "description": "wallet introuvable",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "422": {
                        "description": "le wallet n'est pas actif",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "commun.ErreurDTO": {
            "type": "object",
            "properties": {
                "erreur": {
                    "type": "string",
                    "example": "un compte existe déjà avec cet email"
                }
            }
        },
...
        "wallet.TransactionDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-08-05T18:30:00Z"
                },
                "devise": {
                    "type": "string",
                    "example": "XOF"
                },
                "frais_centimes": {
                    "type": "integer",
                    "example": 75
                },
                "id": {
                    "type": "string",
                    "example": "9c1a2b3d-4e5f-6a7b-8c9d-0e1f2a3b4c5d"
                },
                "montant_centimes": {
                    "type": "integer",
                    "example": 5000
                },
                "operateur": {
                    "type": "string",
                    "example": "ORANGE"
                },
                "statut": {
                    "type": "string",
                    "example": "envoyee"
                },
                "type": {
                    "type": "string",
                    "example": "recharge"
                }
            }
        },
        "wallet.WalletDTO": {
            "type": "object",
            "properties": {
                "devise": {
                    "type": "string",
                    "example": "XOF"
                },
                "id": {
                    "type": "string",
                    "example": "7b8c9d0e-1f2a-4b3c-9d4e-5f6a7b8c9d0e"
                },
                "solde_disponible_centimes": {
                    "type": "integer",
                    "example": 492500
                },
                "solde_en_attente_centimes": {
                    "type": "integer",
                    "example": 0
                },
                "statut": {
                    "type": "string",
                    "example": "actif"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}